Simple repetitive SAP process
The logic is always the same,  firs of al set the sap conextion, for this the sap Script Gui must be active in the sistem. after the conexion you set the loop, and define de variables (the values that change in every loop) and then you can use pretty much every SAP transaction that you want to automate.